refactor(tests): move test-utils to mocks directory and update imports
Update test files to import generateClues from the new location in __mocks__ directory
Remove the validation for minimum number of clues in profile schema test
Simplify shuffle test assertions to use direct array comparison
