+++ src/__mocks__/test-utils.tsx
@@ -1,8 +1,9 @@
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { type RenderOptions, render } from '@testing-library/react';
 import translations from '../../public/locales/en/translation.json';
-import { TranslateProvider } from '../components/TranslateProvider';
 import { ReducedMotionProvider } from '../components/ReducedMotionProvider';
+import { TranslateProvider } from '../components/TranslateProvider';
+import { DEFAULT_CLUES_PER_PROFILE } from '../lib/constants';
 
 const locale = 'en';
 
@@ -39,3 +40,49 @@ export const customRender = (
 
   return render(ui, { wrapper: Wrapper, ...restOptions });
 };
+
+/**
+ * Generates an array of clues with exactly DEFAULT_CLUES_PER_PROFILE (20) items.
+ *
+ * @param customClues - Optional array of custom clues to include
+ *   - If provided with < 20 items: fills remaining slots with generated clues
+ *   - If provided with >= 20 items: uses first 20 items
+ *   - If not provided: generates 20 default clues
+ * @param label - Optional label prefix for clues (e.g., 'Clue')
+ * @returns Array of exactly 20 clues
+ *
+ * @example
+ * // Use default generated clues
+ * const clues = generateClues();
+ *
+ * // Use custom clues with auto-fill
+ * const clues = generateClues(['My custom clue']);
+ *
+ * // Use custom clues with auto-trim
+ * const clues = generateClues(['Clue 1', 'Clue 2', ..., 'Clue 30']);
+ *
+ * // Use custom clues with auto-fill
+ * const clues = generateClues(['My custom clue'], 'Custom Clue');
+ */
+export function generateClues(customClues?: string[], label?: string): string[] {
+  if (!customClues || customClues.length === 0) {
+    // Generate default clues
+    return Array.from(
+      { length: DEFAULT_CLUES_PER_PROFILE },
+      (_, i) => `${label || 'Clue'} ${i + 1}`
+    );
+  }
+
+  if (customClues.length >= DEFAULT_CLUES_PER_PROFILE) {
+    // Trim to exact count if more than needed
+    return customClues.slice(0, DEFAULT_CLUES_PER_PROFILE);
+  }
+
+  // Fill remaining slots with generated clues
+  const result = [...customClues];
+  const customCount = customClues.length;
+  for (let i = customCount; i < DEFAULT_CLUES_PER_PROFILE; i++) {
+    result.push(`Clue ${i + 1}`);
+  }
+  return result;
+}
