+++ src/lib/__tests__/profileLoading.test.ts
@@ -1,8 +1,8 @@
-import { describe, expect, it, vi, beforeEach, afterEach } from 'vitest';
-import { loadProfilesByIds } from '../profileLoading';
-import { generateClues } from '@/__tests__/test-utils';
-import type { Manifest } from '../manifest';
+import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
+import { generateClues } from '@/__mocks__/test-utils';
 import { queryClient } from '../../components/QueryProvider';
+import type { Manifest } from '../manifest';
+import { loadProfilesByIds } from '../profileLoading';
 
 // Mock manifest data
 const mockManifest: Manifest = {
