+++ src/services/__tests__/GamePersistenceService.test.ts
@@ -1,6 +1,6 @@
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
+import { generateClues } from '@/__mocks__/test-utils';
 import { DEFAULT_CLUES_PER_PROFILE } from '@/lib/constants';
-import { generateClues } from '@/__tests__/test-utils';
 import type { PersistedGameState } from '@/lib/gameSessionDB';
 import type { IGameSessionRepository } from '@/repositories/GameSessionRepository';
 import { GamePersistenceService } from '../GamePersistenceService';
